third_party/WebKit/LayoutTests/inspector-protocol/css/css-get-rule-list-expected.txt

Class Selector

ID Selector


Running test: testGetRuleListAfterStartingRecording
=== Size of array: 6
   Number of used Rules: 5
0
1
2
5
16
   Number of unused Rules: 1
3
